Require Jenkins 2.479.3 or newer
Jenkins 2.479.1 is the first LTS version to require Java 17 or newer.
Use 2.479.3 instead of 2.479.1 based on the recommendation in
https://www.jenkins.io/doc/developer/plugin-development/choosing-jenkins-baseline/

https://www.jenkins.io/doc/developer/plugin-development/choosing-jenkins-baseline/
notes that 2.479.1 is the oldest LTS version currently supported by
the Jenkins update center.  It recommends 2.504.3 and 2.516.3 as good
core dependencies.  I didn't update to those versions on the assumption
that testing with Java 25 was more likely to be accepted if fewer changes
were made.